+++ src/components/Inputs/NumberField/NumberField.stories.tsx
@@ -1,0 +1,34 @@
+import { Meta, StoryObj } from "@storybook/react";
+
+import { inputArgTypes } from "../../../internalUtils/inputArgTypes";
+import NumberField from "./NumberField";
+
+const meta: Meta<typeof NumberField> = {
+  title: "Components/Inputs/NumberField",
+  component: NumberField,
+  argTypes: {
+    ...inputArgTypes,
+    prefix: {
+      control: "text",
+      description: "Optional text displayed before value in `NumberField`",
+      table: { type: { summary: "string" } },
+    },
+    value: {
+      control: "number",
+      table: {
+        type: {
+          summary: "number",
+        },
+      },
+    },
+  },
+};
+
+export default meta;
+
+type Story = StoryObj<typeof NumberField>;
+
+export const Default: Story = {};
+export const Label: Story = { args: { label: "Input label", labelType: "left" } };
+export const Size: Story = { args: { size: "small" } };
+export const Error: Story = { args: { error: "Input error" } };