+++ src/components/Inputs/NumberField/helpers/formatValue.ts
@@ -1,0 +1,7 @@
+export const parseNumberToString = (val: number | null): string => {
+  if (val === null) {
+    return "";
+  }
+
+  return val.toString();
+};